gateway/requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
python-dotenv==1.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.7